% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{as.numeric.factor}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{Vector that contains a factor.}
}
\value{
Vector that contais numerics.
}
\description{
Returns a numerical vector that was a factor vector. Is used to create means from likert data.
}
\examples{
x <- factor(c("a", "b", "a"))
as_numeric_factor(x)
}
