% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertplots.R
\name{plikert}
\alias{plikert}
\title{Method to generate a Likert plot using the likert pkg}
\usage{
plikert(items, item_labels = NULL, response_label = "Response",
  xlab = NULL, ylab = "Percent", title = NULL, subtitle = NULL,
  caption = NULL)
}
\arguments{
\item{items}{A data-frame or tibble with ordered factors only.}

\item{item_labels}{A character vector to replace the variable names in the plot}

\item{response_label}{The label for the legend}

\item{xlab}{Label for x-Axis}

\item{ylab}{Label for y-Axis}

\item{title}{A plot title}

\item{subtitle}{A plot subtitle}

\item{caption}{A plot caption}
}
\description{
Method to generate a Likert plot using the likert pkg
}
\examples{
data <- data.frame(a = factor(c("a","b","a")), b=factor(c("a", "b", "b")))
plikert(data)
}
